Added names for StdDev and QC for Leah
@lhendey you can look for each point where there's the line: /**Leah -
Check that "wgtfreq" and "mean" match that of the base file**/

There's one file with the suffix _stdincl for every file that you
created with the original program, with the StdDev/StdErr added as
needed. In them, there should be either a var called "wgtfreq" that
should match the frequency of the basefile for each observation or
"mean", which naturally should match the percentages of the basefile for
each observation. That should be an easy check.